# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MVP multi-tenant event platform. Auth owns user identity; org membership and permissions live in organization/iam.",
        "title": "Rallya API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/events": {
            "get": {
                "description": "Public discovery across orgs, filterable and paginated.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "List published events",
                "parameters": [
                    {
                        "type": "string",
                        "example": "acme",
                        "description": "Org slug filter",
                        "name": "org",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "PUBLISHED",
                        "description": "Status filter",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2026-01-01T00:00:00Z",
                        "description": "RFC3339 start",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2026-12-31T23:59:59Z",
                        "description": "RFC3339 end",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "fest",
                        "description": "Title search",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "starts",
                        "description": "starts|created",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Page (1-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 20,
                        "description": "Page size, max 100",
                        "name": "perPage",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/eventdto.EventsPage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/eventdto.ErrorAlias"
                        }
                    }
                }
            }
        },
        "/events/{id}": {
            "get": {
                "description": "Public detail by UUID. Drafts/cancelled return 404.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get published event",
                "parameters": [
                    {
                        "type": "string",
                        "example": "4d5e6f70-8192-0a1b-2c3d-4e5f6a7b8c9d",
                        "description": "Event UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/eventdto.Event"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/eventdto.ErrorAlias"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/eventdto.ErrorAlias"
                        }
                    }
                }
            }
        },
...
        "/orgs/{id}/events": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Paginated, filterable; includes drafts for members.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "List org events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Org UUID or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "PUBLISHED",
                        "description": "DRAFT/PUBLISHED/CANCELLED",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2026-01-01T00:00:00Z",
                        "description": "RFC3339 start",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2026-12-31T23:59:59Z",
                        "description": "RFC3339 end",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "fest",
                        "description": "Title search",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "starts",
                        "description": "starts|created",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Page (1-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 20,
                        "description": "Page size, max 100",
                        "name": "perPage",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/eventdto.EventsPage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/eventdto.ErrorAlias"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/eventdto.ErrorAlias"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/eventdto.ErrorAlias"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a DRAFT event under the org; slug auto-uniquifies per org.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Create event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Org UUID or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Event payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/eventdto.CreateEvent"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/eventdto.Event"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/eventdto.ErrorAlias"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/eventdto.ErrorAlias"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/eventdto.ErrorAlias"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/eventdto.ErrorAlias"
                        }
                    }
                }
            }
        },
...
        "/orgs/{id}/events/{eventId}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "PUBLISHED → CANCELLED (terminal).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Cancel event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Org UUID or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Event UUID or slug",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/eventdto.Event"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/eventdto.ErrorAlias"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/eventdto.ErrorAlias"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/eventdto.ErrorAlias"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/eventdto.ErrorAlias"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "eventdto.ErrorAlias": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "event not found"
                }
            }
        },
        "eventdto.Event": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer",
                    "example": 500
                },
                "coverUrl": {
                    "type": "string",
                    "example": "/uploads/events/acme/x7f2k9.jpg"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2026-09-19T12:00:00Z"
                },
                "description": {
                    "type": "string"
                },
                "endsAt": {
                    "type": "string",
                    "example": "2026-08-03T22:00:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "4d5e6f70-8192-0a1b-2c3d-4e5f6a7b8c9d"
                },
                "location": {
                    "type": "string"
                },
                "slug": {
                    "type": "string",
                    "example": "summer-fest-2026"
                },
                "startsAt": {
                    "type": "string",
                    "example": "2026-08-01T10:00:00Z"
                },
                "status": {
                    "type": "string",
                    "example": "PUBLISHED"
                },
                "title": {
                    "type": "string",
                    "example": "Summer Fest 2026"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2026-09-19T12:00:00Z"
                },
                "venue": {
                    "type": "string"
                }
            }
        },
...
        "eventdto.EventsPage": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/eventdto.Event"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "perPage": {
                    "type": "integer",
                    "example": 20
                },
                "total": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer \u003caccessToken\u003e from POST /auth/login (or /auth/refresh).\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}